backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "WonderComic backend - AI-powered comic book stories"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.2.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "google-genai>=1.0.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "aiofiles>=25.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-mock>=3.12.0",
    "requests>=2.32.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
